Criação da pagina de funcionarios
criação do front e back de funcionarios

+++ GestaoEscalaPermuta.Infra/Data/EntitiesDefesaCivilMarica/Funcionario.cs
@@ -18,7 +18,7 @@ public partial class Funcionario
 
     public int NrMatricula { get; set; }
 
-    public int? NrTelefone { get; set; }
+    public long? NrTelefone { get; set; }
 
     [StringLength(200)]
     [Unicode(false)]
